% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCTDRosette.R
\name{getCTDRosette}
\alias{getCTDRosette}
\title{CTD Rosette Operations}
\usage{
getCTDRosette(metric = "NULL", expedition = "NULL")
}
\arguments{
\item{metric}{A string indicating the metric.
\itemize{
\item{Choose from one of the following: }{"conductivity", "temperature", "depth", "salinity", "oxygen",
"oxidation reduction potential", "turbidity", "fluorometer"  or "All"}
\item{Default value: }{"All"}
\item{Note not all metrics were collected on each expedition.}
}}

\item{expedition}{A string indicating the expedition.
\itemize{
\item{Choose from one of the following: }
\itemize{
\item "EX1805" (East Florida Telemapping),
\item "EX1810" (Mapping Deepwater Areas in the Caribbean and South Atlantic Bight),
\item "EX1812" (Caribbean/SAB ROV and Mapping),
\item "EX1903" (Southeastern US Atlantic Continental Margin Mapping),
\item "EX1905" (New England and Canada Mapping),
\item "EX1906" (Southeast US and Bahamas, Leg 1 Mapping),
\item "EX2101" (2021 EM304 SAT + Mapping Shakedown),
\item "EX2102" (2021 Technology Demonstration),
\item "EX2106" (2021 U.S. Blake Plateau Mapping 2),
\item "EX2107" (Windows to the Deep 2021: Blake Plateau),
\item "EX2202" (Caribbean Mapping ),
\item "EX2203" (2022 Puerto Rico Mapping and Deep-Sea Camera Demonstration),
\item Or you can select "All" for all expeditions.
}

\item{Default value = }{"All"}
\item{See https://oceanexplorer.noaa.gov/explorations/aspire/welcome.html
for a list of all ASPIRE expeditions}
}}
}
\value{
This function returns a dataframe with the following columns:
\describe{
\item{YEAR}{The year}
\item{EXPEDITION}{A code indicating a specified cruise, ex: EX2203 or all expeditions, ex. "ALL"}
\item{FEATURE}{A code for the geological feature surveyed. If "All" expeditions is specified, FEATURE will = "ALL"}
\item{METRIC}{Possible metrics:}
\itemize{
\item{conductivity (S/m)},
\item{temperature (C)}
\item{depth (m)}
\item{salinity}
\item{oxygen (\%)}
\item{oxygen2 (mg/l)}
\item{oxygen3 (μmol/kg)}
\item{density (kg/m³)}
\item{sound speed (m/s)}
\item{pressure (dbar)}
}
}
}
\description{
Creates CTD Rosette dataframes based on specified metric or expedition
}
\details{
NOAA Ocean Exploration collects oceanographic data and water samples by request during mapping and ROV
expeditions on \emph{Okeanos Explorer} with a CTD rosette system. The CTD rosette contains conductivity,
temperature, depth, oxygen, salinity, oxidation reduction potential, turbidity, and fluorometer sensors.
CTD rosette data are archived as .hex and .cnv files, which are produced using Seabird software.The oce R package
(Kelley & Richards 2020) was used to read in the .cnv file into R where the CTD rosette data were extracted.
}
\examples{
#Get temperature profiles from all ASPRIRE cruises
EX_CTD <- getCTDRosette(metric = "temperature", expedition = "All")
#Returns all of the CTD Rosette temperature data collected during the ASPIRE campaign

#Get all of the CTD Rosette data collected from a single expedition
EX_CTD <- getCTDRosette(metric = "ALL", expedition = "EX1903")
#Returns all of the CTD Rosette data collected during the EX1903 expedition
}
\seealso{
\code{\link{getROVCTD}}
}
